requirements.txt
requests
beautifulsoup4
qrcode[pil]
Pillow 